refactor(#1519): move memory_router + relationship_extractor from core/ to services/memory/
These files are service-layer code that was incorrectly placed in core/:
- memory_router.py (820-line MemoryViewRouter) → services/memory/
- relationship_extractor.py (NLP extractors) → services/memory/

Also fixes 4 stale imports in nexus_fs.py that still pointed to
nexus.core.memory_api / nexus.core.memory_with_paging (moved earlier
but imports were missed).

Updated 21 import statements across 11 source + 4 test files.
No backward compat shims per PM decision.
